fix: resolve CI backend-size and UI coverage merge failures for PR #625
Split cron workstation mapper tests into a dedicated file under the 1000-line
limit and extend add-dialog coverage for LOGICAL_MOVE type switching and cron
field edits so merged UI coverage clears the global threshold.
